Refactored Modal code
